Backend Calculator API Project  - Adapted to the usage of float32 type  - Added RequestId through a Middleware

+++ backend-api-calculator/calculator/model.go
@@ -1,15 +1,15 @@
 package calculator
 
 type TwoNumbersRequest struct {
-	Number1 int `json:"number1"`
-	Number2 int `json:"number2"`
+	Number1 float32 `json:"number1"`
+	Number2 float32 `json:"number2"`
 }
 
 type DivideRequest struct {
-	Dividend int `json:"dividend"`
-	Divisor  int `json:"divisor"`
+	Dividend float32 `json:"dividend"`
+	Divisor  float32 `json:"divisor"`
 }
 
 type ResultResponse struct {
-	Result int `json:"result"`
+	Result float32 `json:"result"`
 }
